feat(submissions): move filter logic to submission util
This function does not call repositories or other services, and
behaves like a standalone block of business logic. It might be more
suited to be placed in a util file. Also, because it is used by both
encrypt and email submissions, I'll put it in the parent directory util.
